requirements.txt
Django
djangorestframework
psycopg2-binary
gunicorn
drf-yasg
django-filter

reportlab
